Fix some wrong conversion types introduced in/missed by #1709

+++ src/main/java/moze_intel/projecte/gameObjs/container/slots/transmutation/SlotConsume.java
@@ -28,7 +28,7 @@ public class SlotConsume extends SlotItemHandler
 		
 		ItemStack cache = stack.copy();
 		
-		double toAdd = 0;
+		long toAdd = 0;
 		
 		while (!inv.hasMaxedEmc() && stack.getCount() > 0)
 		{
